+++ src/gui/addCrystalPopulationButton.cpp
@@ -7,11 +7,11 @@ AddCrystalPopulationButton::AddCrystalPopulationButton(QWidget *parent)
 
     mMenu = new QMenu(this);
 
-    mAddRandom = new QAction("Add random", this);
-    mAddPlate = new QAction("Add plate", this);
-    mAddColumn = new QAction("Add column", this);
-    mAddParry = new QAction("Add Parry", this);
-    mAddLowitz = new QAction("Add Lowitz", this);
+    mAddRandom = new QAction(tr("Add random"), this);
+    mAddPlate = new QAction(tr("Add plate"), this);
+    mAddColumn = new QAction(tr("Add column"), this);
+    mAddParry = new QAction(tr("Add Parry"), this);
+    mAddLowitz = new QAction(tr("Add Lowitz"), this);
 
     mMenu->addActions({mAddRandom,
                        mAddPlate,